requirements.txt
msal
requests
inquirer
tqdm
python-dotenv
